requirements.txt
gradio>=3.41
Pillow
numpy
scikit-learn